Create filter component for search
